# account/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "docker监控服务后端API接口文档",
        "title": "Docker监控服务",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8889",
    "basePath": "/api",
    "paths": {
        "/role/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "添加角色",
                "parameters": [
                    {
                        "description": "roleModel",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.AddRoleModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回空对象",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/role/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "角色列表",
                "responses": {
                    "200": {
                        "description": "返回空对象",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "添加用户",
                "responses": {
                    "200": {
                        "description": "返回空对象",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/user/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "添加用户",
                "parameters": [
                    {
                        "description": "修改用户",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.EditUserModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回空对象",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "登入",
                "parameters": [
                    {
                        "description": "登入",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.LoginModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回空对象",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "注册",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.RegisterModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回空对象",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "添加用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回空对象",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "service.AddRoleModel": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "角色code",
                    "type": "string"
                },
                "name": {
                    "description": "角色名",
                    "type": "string"
                }
            }
        },
        "service.AddUserModel": {
            "type": "object",
            "properties": {
                "nickname": {
                    "description": "用户昵称",
                    "type": "string"
                },
                "password": {
                    "description": "用户登录密码",
                    "type": "string"
                },
                "roles": {
                    "description": "所属角色",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "description": "用户登录名",
                    "type": "string"
                }
            }
        },
        "service.EditUserModel": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "用户id",
                    "type": "integer"
                },
                "nickname": {
                    "description": "用户昵称",
                    "type": "string"
                },
                "password": {
                    "description": "用户登录密码",
                    "type": "string"
                },
                "roles": {
                    "description": "所属角色",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "description": "用户登录名",
                    "type": "string"
                }
            }
        },
        "service.LoginModel": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "service.RegisterModel": {
            "type": "object",
            "properties": {
                "nickname": {
                    "description": "用户昵称",
                    "type": "string"
                },
                "password": {
                    "description": "用户登录密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户登录名",
                    "type": "string"
                }
            }
        }
    }
}